requirements.txt
pydantic
duckduckgo-search
bs4
requests
rich
python-dotenv
openai-agents